requirements.txt
flask_restplus
flask
Pillow
opencv-python
numpy